backend/requirements.txt
Django>=5.1.4
requests
python-dotenv
openai
